:hammer: refactor: 공통 DTO 파일 이동
- 기존의 commonDTO 패키지에서 common으로 이동
